third_party/blink/web_tests/external/wpt/service-workers/service-worker/local-url-inherit-controller.https-expected.txt

This is a testharness.js-based test.
[FAIL] Same-origin blob URL iframe should inherit service worker controller.
  assert_equals: blob URL iframe should inherit controller expected (string) "https://web-platform.test:8444/service-workers/service-worker/resources/local-url-inherit-controller-worker.js" but got (object) null
[FAIL] Same-origin blob URL iframe should intercept fetch().
  assert_equals: blob URL iframe should intercept fetch expected "intercepted" but got "var hello = \\"world\\";\\n"
[FAIL] Same-origin blob URL worker should inherit service worker controller.
  promise_test: Unhandled rejection with value: "Uncaught TypeError: Cannot read properties of undefined (reading 'controller')"
[FAIL] Same-origin blob URL sharedworker should inherit service worker controller.
  promise_test: Unhandled rejection with value: "Uncaught TypeError: Cannot read properties of undefined (reading 'controller')"
[FAIL] Data URL worker should not inherit service worker controller.
  promise_test: Unhandled rejection with value: "Uncaught TypeError: Cannot read properties of undefined (reading 'controller')"
Harness: the test ran to completion.
